Added base vue from https://codepen.io/CSWApps/pen/oWJZqQ
This is code to autocomplete a field and get the values needed. I made it so there were two but I am unsure of effiency for my purposes. I am going to read up on how to best use it with two fields as I currently have never worked with Vue. Will continue to figure out especially once I have access to a wifi instead of just a hotspot.